Reference

Schedulers.epmap — Method

epmap(f, tasks, args...; pmap_kwargs..., f_kwargs...)

where f is the map function, and tasks is an iterable collection of tasks. The function f takes the positional arguments args, and the keyword arguments f_args. The optional arguments pmap_kwargs are as follows.

pmap_kwargs

epmap_retries=0 number of times to retry a task on a given machine before removing that machine from the cluster

epmap_maxerrors=Inf the maximum number of errors before we give-up and exit

epmap_minworkers=nworkers() the minimum number of workers to elastically shrink to

epmap_maxworkers=nworkers() the maximum number of workers to elastically expand to

epmap_usemaster=false assign tasks to the master process?

epmap_nworkers=nworkers the number of machines currently provisioned for work[1]

epmap_quantum=32 the maximum number of workers to elastically add at a time

epmap_addprocs=n->addprocs(n) method for adding n processes (will depend on the cluster manager being used)

epmap_init=pid->nothing after starting a worker, this method is run on that worker.

epmap_preempted=()->false method for determining of a machine got pre-empted (removed on purpose)[2]

epmap_reporttasks=true log task assignment

Notes

[1] The number of machines provisioined may be greater than the number of workers in the cluster since with some cluster managers, there may be a delay between the provisioining of a machine, and when it is added to the Julia cluster. [2] For example, on Azure Cloud a SPOT instance will be pre-empted if someone is willing to pay more for it

Schedulers.epmapreduce! — Method

epmapreduce!(result, f, tasks, args...; epmap_kwargs..., kwargs...) -> result

where f is the map function, and tasks are an iterable set of tasks to map over. The positional arguments args and the named arguments kwargs are passed to f which has the method signature: f(localresult, f, task, args; kwargs...). localresult is a Future with an assoicated partial reduction.

epmap_kwargs

epmap_reducer! = default_reducer! the method used to reduce the result. The default is (x,y)->(x .+= y)

epmap_zeros = ()->zeros(eltype(result), size(result)) the method used to initiaize partial reductions

epmap_minworkers=nworkers() the minimum number of workers to elastically shrink to

epmap_maxworkers=nworkers() the maximum number of workers to elastically expand to

epmap_usemaster=false assign tasks to the master process?

epmap_nworkers=nworkers the number of machines currently provisioned for work[1]

epmap_quantum=32 the maximum number of workers to elastically add at a time

epmap_addprocs=n->addprocs(n) method for adding n processes (will depend on the cluster manager being used)

epmap_init=pid->nothing after starting a worker, this method is run on that worker.

epmap_scratch="/scratch" storage location accesible to all cluster machines (e.g NFS, Azure blobstore,...)

epmap_reporttasks=true log task assignment

Notes

[1] The number of machines provisioined may be greater than the number of workers in the cluster since with some cluster managers, there may be a delay between the provisioining of a machine, and when it is added to the Julia cluster. [2] For example, on Azure Cloud a SPOT instance will be pre-empted if someone is willing to pay more for it

Examples

Example 1

With the assumption that /scratch is accesible from all workers:

using Distributed
addprocs(2)
@everywhere using Distributed, Schedulers
@everywhere f(x, tsk) = (fetch(x)::Vector{Float32} .+= tsk; nothing)
result = epmapreduce!(zeros(Float32,10), f, 1:100)
rmprocs(workers())

Example 2

Using Azure blob storage:

using Distributed, AzStorage
container = AzContainer("scratch"; storageaccount="mystorageaccount")
mkpath(container)
addprocs(2)
@everywhere using Distributed, Schedulers
@everywhere f(x, tsk) = (fetch(x)::Vector{Float32} .+= tsk; nothing)
result = epmapreduce!(zeros(Float32,10), f, 1:100; epmap_scratch=container)
rmprocs(workers())